+++ src/main/java/com/erp/system/hr/model/attendance_information_management/Attendance.java
@@ -1,4 +1,4 @@
-package com.erp.system.hr.Model.attendance_information_management;
+package com.erp.system.hr.model.attendance_information_management;
 
 import com.erp.system.common.annotation.EnumMapping;
 import com.erp.system.common.validator.NotNullValidator;
@@ -7,7 +7,6 @@ import com.erp.system.common.annotation.Unique;
 
 import java.sql.Time;
 import java.time.LocalDate;
-import java.util.Date;
 
 
 // 근태 기록 등록 테이블
