+++ src/main/java/com/erp/system/hr/model/salary_information_management/Salary.java
@@ -1,4 +1,4 @@
-package com.erp.system.hr.Model.salary_information_management;
+package com.erp.system.hr.model.salary_information_management;
 
 import com.erp.system.common.validator.NotNullValidator;
 import com.erp.system.common.annotation.NotNull;
