+++ src/main/java/com/erp/system/hr/model/salary_information_management/SalaryType.java
@@ -1,11 +1,8 @@
-package com.erp.system.hr.Model.salary_information_management;
+package com.erp.system.hr.model.salary_information_management;
 
 import com.erp.system.common.annotation.NotNull;
 import com.erp.system.common.annotation.Unique;
 import com.erp.system.common.validator.NotNullValidator;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
 
 // 급여 유형 정보 테이블
