+++ src/main/java/com/erp/system/hr/respository/UserRepository.java
@@ -1,9 +1,7 @@
 package com.erp.system.hr.respository;
 
-import com.erp.system.common.annotation.Component;
 import com.erp.system.common.generic_repository.GenericRepository;
-import com.erp.system.financial.model.book_keeping.accounting_ledger.CashBook;
-import com.erp.system.hr.Model.basic_information_management.User;
+import com.erp.system.hr.model.basic_information_management.User;
 
 import java.util.List;
