+++ src/main/java/com/erp/system/hr/respository/UserRepositoryImpl.java
@@ -2,9 +2,7 @@ package com.erp.system.hr.respository;
 
 import com.erp.system.common.annotation.Component;
 import com.erp.system.common.generic_repository.impl.GenericRepositoryImpl;
-import com.erp.system.financial.model.book_keeping.accounting_ledger.GeneralLedger;
-import com.erp.system.financial.repository.book_keeping.accounting_ledger.GeneralLedgerRepository;
-import com.erp.system.hr.Model.basic_information_management.User;
+import com.erp.system.hr.model.basic_information_management.User;
 
 @Component
 public class UserRepositoryImpl extends GenericRepositoryImpl<User> implements UserRepository {
