+++ src/main/java/com/erp/system/hr/service/basic_information_management/UserService.java
@@ -1,7 +1,7 @@
 package com.erp.system.hr.service.basic_information_management;
 
-import com.erp.system.hr.Model.basic_information_management.User;
-import com.erp.system.hr.Model.dto.UserDto;
+import com.erp.system.hr.model.basic_information_management.User;
+import com.erp.system.hr.model.dto.UserDto;
 
 public interface  UserService {
     void User(UserDto dto);
